Move HttpClient to core directory
